requirements.txt
django~=5.1.3
psutil~=6.1.0
drf-yasg~=1.21.8
mysqlclient~=2.2.6
python-decouple~=3.8
django-filter~=2.4.0
django-debug-toolbar~=4.4.6
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
better-profanity==0.7.0